+++ src/components/radio-group/radio-group.test.ts
@@ -297,6 +297,102 @@ describe('<wa-radio-group>', () => {
       });
     });
 
+    describe('when handling focus', () => {
+      const doAction = async (instance: WaRadioGroup, type: string) => {
+        if (type === 'focus') {
+          instance.focus();
+          await instance.updateComplete;
+          return;
+        }
+
+        const label = instance.shadowRoot!.querySelector<HTMLLabelElement>('#label')!;
+        label.click();
+        await instance.updateComplete;
+      };
+
+      // Tests for focus and label actions with radio buttons
+      ['focus', 'label'].forEach(actionType => {
+        describe(`when using ${actionType}`, () => {
+          it('should do nothing if all elements are disabled', async () => {
+            const el = await fixture<WaRadioGroup>(html`
+              <wa-radio-group>
+                <wa-radio id="radio-0" value="0" disabled></wa-radio>
+                <wa-radio id="radio-1" value="1" disabled></wa-radio>
+                <wa-radio id="radio-2" value="2" disabled></wa-radio>
+                <wa-radio id="radio-3" value="3" disabled></wa-radio>
+              </wa-radio-group>
+            `);
+
+            const validFocusHandler = sinon.spy();
+
+            Array.from(el.querySelectorAll<WaRadio>('wa-radio')).forEach(radio =>
+              radio.addEventListener('wa-focus', validFocusHandler)
+            );
+
+            expect(validFocusHandler).to.not.have.been.called;
+            await doAction(el, actionType);
+            expect(validFocusHandler).to.not.have.been.called;
+          });
+
+          it('should focus the first radio that is enabled when the group receives focus', async () => {
+            const el = await fixture<WaRadioGroup>(html`
+              <wa-radio-group>
+                <wa-radio id="radio-0" value="0" disabled></wa-radio>
+                <wa-radio id="radio-1" value="1"></wa-radio>
+                <wa-radio id="radio-2" value="2"></wa-radio>
+                <wa-radio id="radio-3" value="3"></wa-radio>
+              </wa-radio-group>
+            `);
+
+            const invalidFocusHandler = sinon.spy();
+            const validFocusHandler = sinon.spy();
+
+            const disabledRadio = el.querySelector('#radio-0')!;
+            const validRadio = el.querySelector('#radio-1')!;
+
+            disabledRadio.addEventListener('wa-focus', invalidFocusHandler);
+            validRadio.addEventListener('wa-focus', validFocusHandler);
+
+            expect(invalidFocusHandler).to.not.have.been.called;
+            expect(validFocusHandler).to.not.have.been.called;
+
+            await doAction(el, actionType);
+
+            expect(invalidFocusHandler).to.not.have.been.called;
+            expect(validFocusHandler).to.have.been.called;
+          });
+
+          it('should focus the currently enabled radio when the group receives focus', async () => {
+            const el = await fixture<WaRadioGroup>(html`
+              <wa-radio-group value="2">
+                <wa-radio id="radio-0" value="0" disabled></wa-radio>
+                <wa-radio id="radio-1" value="1"></wa-radio>
+                <wa-radio id="radio-2" value="2" checked></wa-radio>
+                <wa-radio id="radio-3" value="3"></wa-radio>
+              </wa-radio-group>
+            `);
+
+            const invalidFocusHandler = sinon.spy();
+            const validFocusHandler = sinon.spy();
+
+            const disabledRadio = el.querySelector('#radio-0')!;
+            const validRadio = el.querySelector('#radio-2')!;
+
+            disabledRadio.addEventListener('wa-focus', invalidFocusHandler);
+            validRadio.addEventListener('wa-focus', validFocusHandler);
+
+            expect(invalidFocusHandler).to.not.have.been.called;
+            expect(validFocusHandler).to.not.have.been.called;
+
+            await doAction(el, actionType);
+
+            expect(invalidFocusHandler).to.not.have.been.called;
+            expect(validFocusHandler).to.have.been.called;
+          });
+        });
+      });
+    });
+
     describe('when the value changes', () => {
       it('should emit wa-change when toggled with the arrow keys', async () => {
         const radioGroup = await fixture<WaRadioGroup>(html`